+++ Stripe/PaymentSheet.swift
@@ -176,12 +176,24 @@ public class PaymentSheet {
         presentingViewController.presentAsBottomSheet(bottomSheetViewController, appearance: configuration.appearance)
     }
 
-    /// Deletes all persisted state.
+    /// Deletes all persisted authentication state associated with a customer.
     ///
     /// You must call this method when the user logs out from your app.
-    /// This will ensure that any persisted state in PaymentSheet, such as
-    /// authentication cookies, is also cleared during logout.
+    /// This will ensure that any persisted authentication state in PaymentSheet,
+    /// such as authentication cookies, is also cleared during logout.
+    ///
+    /// - Warning: Deprecated. Use `PaymentSheet.resetCustomer()` instead.
+    @available(*, deprecated, renamed: "resetCustomer()")
     public static func reset() {
+        resetCustomer()
+    }
+
+    /// Deletes all persisted authentication state associated with a customer.
+    ///
+    /// You must call this method when the user logs out from your app.
+    /// This will ensure that any persisted authentication state in PaymentSheet,
+    /// such as authentication cookies, is also cleared during logout.
+    public static func resetCustomer() {
         LinkAccountService.defaultCookieStore.clear()
     }